Move codec utils out of mse dir and add tests
